remove cubdb and use cachex
Since Koyeb loses the data file when the application is put to sleep there is no need to use cubdb for now.
Pocket caches the article.
Once it is added, it requests it and beyond that we don't need to store it. It could be removed right away but this is more forgiving and Koyeb will reset the state anyway or the expiration kicks in.